xraysearch-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
hnswlib==0.8.0
numpy==1.26.4
Pillow==10.3.0
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
pandas==2.2.2
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
